examples/python/image_generation/requirements.txt
cakework
banana_dev
pillow
boto3
nanoid